Update Introductory Database Tutorial
In the Workshop section at the end of the notebook, I added code to connect to the sqlite database. There had been some confusion during class because the instructions only tell students to insert contacts into the database and print out a list of companies and their associated contacts. In addition, the original layout of the template students are asked to use didn't include a space for the connection or a prompt reminding them that it needed to created, which led many students to miss this step during the inclass workshop.